move isSea type guard to navigation folder
