+++ apps/bot/src/domains/event/generators/passive/rough-sea.ts
@@ -1,6 +1,7 @@
 import { buildOpEmbed } from '../../../../discord/utils/build-op-embed.js';
+import { isSea } from '../../../navigation/utils/index.js';
 import type { PassiveGenerator } from '../../types.js';
-import { computeNothing, isSea } from '../utils.js';
+import { computeNothing } from '../utils.js';
 
 export const roughSea: PassiveGenerator = {
   key: 'roughSea',
