+++ apps/bot/src/domains/event/generators/utils.ts
@@ -1,13 +1,7 @@
-import { SEAS, type Sea, type Zone } from '@one-piece/db';
-
 import type { Rng } from '../types.js';
 
 export function computeNothing() {
   return { effects: [], state: {} };
-}
-
-export function isSea(zone: Zone): zone is Sea {
-  return (SEAS as ReadonlyArray<Zone>).includes(zone);
 }
 
 /** Entier dans [min, max] (bornes incluses) tiré depuis le rng fourni. */
