+++ apps/bot/src/domains/navigation/utils/index.ts
@@ -1,3 +1,4 @@
+export { isSea } from './is-sea.js';
 export { computeDriftProbability } from './compute-drift-probability.js';
 export { findEdge } from './find-edge.js';
 export { hasLogOrEternalPoseForIsland } from './has-log-or-eternal-pose-for-island.js';
